feat(cooldown-hook): Dirty rotten PoC
Added part of Caraxi's code, especially the ActionManager which can be
queried to get a Cooldown struct for a given CooldownGroup.
Amazingly, that combined with the CooldownGroup of an action (taken from
Excel Sheet via Lumina) is all that is needed for the basic
functionality.

Thanks to Caraxi doing all the exploration for the
Cooldown/Action related stuff.

Now on to making a system for vibration patterns.