feat(hook): Finish exploration
Hook for `XInputWrapper.SetState` working, as is the hook for the
function in FFXIV ('FFXIV.SetState') which calls SetState.

Hook for a possible GameControls(?) Ctor is not working but superseeded
anyways via a hook to a function which polls the active controller. This
function gets the pointer to the GameControls(?) struct which is the
same pointer we need for `FFXIV.SetState`.
As the polling happens all the time, we can grab that instantly and
disable the hook and now it is possible to go BRBR with the controller
without having the game causes vibration first (we did that first to get
the correct pointer, before the PollHook was discovered).

After this exploration and working PoC now comes the phase with UI
design and hooking into Cooldowns/GCD. Thankfully, RemindMe already does
everything we need :D.