feat(config): Add basic step and pattern editing
I needed to work around Imgui and my desire for immutable data.
The resulting abomination is....not pleasant.
I would need to either (a) ditch the immutable config or (b) try if
`with` updates for the record would work. That would create garbage
though but I already creating garbage by doing the hop through MutableXY
so....
For now though, it works and I had enough of that. On to other problems!