feat(priority): Vibrate in importance order WIP
This code stinks. And it is quiestionable if this is even desireable
behaviour. Per virtue of one controller only on vibration (pattern) can
be active at a time. So what to do if during a pattern another should
trigger? Currently, we assign a priority to each trigger and let higher
priority ones interupt the current pattern. But, does this makes sense?
Improves the usefullness of the vibration feature?
Furthermore, we let other triggers queue so that after the current,
highest priority one ends the next one can vibrate. Realisticly though,
that complicates a lot for .... what benefit?