feat(priorityqueue): A somewhat working queue and trigger order
Triggers have a priority, triggers are handled in priority order,
Triggers with inifinite patterns are started again if interrupted,
Lower priority triggers are queued. Is it useful? Who knows, at least it
is a little less shit then it was before.
Had to rip out the newest PriorityQueue proposal for .NET 6 though, as
there is no implementation in current version.

Ah, Configuration is not immutable anymore. I gave up, not enough
experience to create a hassle free facade/experience.
Now I just have to follow the rule of cool:
Changing config outside of ConfigurationUI? Not cool.